egeklfnsld

+++ SI_IFES/PROG_II/ex-presencial/lista1/lista 1/ex_4.c
@@ -16,7 +16,7 @@ valor do desconto. */
 #define CORTE "-------------------------------------"
 
 
-int pedirNum(){
+float pedirNum(){
     float num;
     printf("Digite o preço do produto: ");
     scanf("%f", &num);
